Refactor timestamp handling in AaveYieldAdapter and RestlessEscrow contracts
- Change `depositTimestamp` in `AaveYieldAdapter` and `createdAt`, `fundedAt`, `disputedAt` in `RestlessEscrow` to use `uint40` for improved storage efficiency.
- Update related functions to ensure compatibility with the new timestamp types, enhancing overall contract performance and reducing gas costs.
- Adjust `CreateDealParams` and `Deal` structs in `Types.sol` to reflect the new timeout type as `uint32`, streamlining parameter handling across contracts.

+++ contracts/contracts/AaveYieldAdapter.sol
@@ -19,7 +19,7 @@ contract AaveYieldAdapter is IYieldAdapter {
     struct DepositRecord {
         uint256 principal;
         uint256 aTokenBalance;
-        uint256 depositTimestamp;
+        uint40 depositTimestamp;
         bool active;
     }
 
@@ -80,7 +80,7 @@ contract AaveYieldAdapter is IYieldAdapter {
         deposits[dealId] = DepositRecord({
             principal: amount,
             aTokenBalance: aTokenReceived,
-            depositTimestamp: block.timestamp,
+            depositTimestamp: uint40(block.timestamp),
             active: true
         });
         totalDeposited += amount;
